[HER-53]Follow-up Move files
