+++ src/Data/Detection/DetectionFactory.cs
